backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
mysqlclient>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.5
requests>=2.26.0
python-dotenv>=0.19.0
pydantic>=1.8.2
proxmoxer>=1.3.0
guacli>=0.1.0
pytest>=6.2.5
alembic>=1.7.1